moved generated files to src/openapi
