--- docs/openapi/redoc-specs.js
@@ -1,1 +1,0 @@
-module.exports = [{"spec":"docs/openapi/keymap.yaml","route":"openapi/testkube-keymap"},{"spec":"docs/openapi/triggers.yaml","route":"openapi/testkube-triggers"},{"spec":"docs/openapi/test-suites.yaml","route":"openapi/testkube-test-suites"},{"spec":"docs/openapi/test-suite-with-executions.yaml","route":"openapi/testkube-test-suite-with-executions"},{"spec":"docs/openapi/test-suite-executions.yaml","route":"openapi/testkube-test-suite-executions"},{"spec":"docs/openapi/executions.yaml","route":"openapi/testkube-executions"},{"spec":"docs/openapi/tests.yaml","route":"openapi/testkube-tests"},{"spec":"docs/openapi/test-with-executions.yaml","route":"openapi/testkube-test-with-executions"},{"spec":"docs/openapi/executors.yaml","route":"openapi/testkube-executors"},{"spec":"docs/openapi/executor-by-types.yaml","route":"openapi/testkube-executor-by-types"},{"spec":"docs/openapi/labels.yaml","route":"openapi/testkube-labels"},{"spec":"docs/openapi/webhooks.yaml","route":"openapi/testkube-webhooks"},{"spec":"docs/openapi/templates.yaml","route":"openapi/testkube-templates"},{"spec":"docs/openapi/config.yaml","route":"openapi/testkube-config"},{"spec":"docs/openapi/debug.yaml","route":"openapi/testkube-debug"},{"spec":"docs/openapi/test-sources.yaml","route":"openapi/testkube-test-sources"},{"spec":"docs/openapi/uploads.yaml","route":"openapi/testkube-uploads"},{"spec":"docs/openapi/repositories.yaml","route":"openapi/testkube-repositories"},{"spec":"docs/openapi/secrets.yaml","route":"openapi/testkube-secrets"},{"spec":"docs/openapi/test-workflows.yaml","route":"openapi/testkube-test-workflows"},{"spec":"docs/openapi/test-workflow-with-executions.yaml","route":"openapi/testkube-test-workflow-with-executions"},{"spec":"docs/openapi/test-workflow-executions.yaml","route":"openapi/testkube-test-workflow-executions"},{"spec":"docs/openapi/preview-test-workflow.yaml","route":"openapi/testkube-preview-test-workflow"},{"spec":"docs/openapi/test-workflow-templates.yaml","route":"openapi/testkube-test-workflow-templates"}]
