+++ src/openapi/redoc-sidebar.js
