+++ src/openapi/redoc-specs.js
@@ -1,0 +1,1 @@
+module.exports = [{"spec":"src/openapi/keymap.yaml","route":"openapi/testkube-keymap"},{"spec":"src/openapi/triggers.yaml","route":"openapi/testkube-triggers"},{"spec":"src/openapi/test-suites.yaml","route":"openapi/testkube-test-suites"},{"spec":"src/openapi/test-suite-with-executions.yaml","route":"openapi/testkube-test-suite-with-executions"},{"spec":"src/openapi/test-suite-executions.yaml","route":"openapi/testkube-test-suite-executions"},{"spec":"src/openapi/executions.yaml","route":"openapi/testkube-executions"},{"spec":"src/openapi/tests.yaml","route":"openapi/testkube-tests"},{"spec":"src/openapi/test-with-executions.yaml","route":"openapi/testkube-test-with-executions"},{"spec":"src/openapi/executors.yaml","route":"openapi/testkube-executors"},{"spec":"src/openapi/executor-by-types.yaml","route":"openapi/testkube-executor-by-types"},{"spec":"src/openapi/labels.yaml","route":"openapi/testkube-labels"},{"spec":"src/openapi/webhooks.yaml","route":"openapi/testkube-webhooks"},{"spec":"src/openapi/templates.yaml","route":"openapi/testkube-templates"},{"spec":"src/openapi/config.yaml","route":"openapi/testkube-config"},{"spec":"src/openapi/debug.yaml","route":"openapi/testkube-debug"},{"spec":"src/openapi/test-sources.yaml","route":"openapi/testkube-test-sources"},{"spec":"src/openapi/uploads.yaml","route":"openapi/testkube-uploads"},{"spec":"src/openapi/repositories.yaml","route":"openapi/testkube-repositories"},{"spec":"src/openapi/secrets.yaml","route":"openapi/testkube-secrets"},{"spec":"src/openapi/test-workflows.yaml","route":"openapi/testkube-test-workflows"},{"spec":"src/openapi/test-workflow-with-executions.yaml","route":"openapi/testkube-test-workflow-with-executions"},{"spec":"src/openapi/test-workflow-executions.yaml","route":"openapi/testkube-test-workflow-executions"},{"spec":"src/openapi/preview-test-workflow.yaml","route":"openapi/testkube-preview-test-workflow"},{"spec":"src/openapi/test-workflow-templates.yaml","route":"openapi/testkube-test-workflow-templates"}]
