+++ src/openapi/templates.yaml
