+++ src/openapi/uploads.yaml
